refactor: move DTO classes to application packages
