+++ src/main/java/kitchenpos/menu/application/dto/MenuCreationRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.ui.dto;
+package kitchenpos.menu.application.dto;
 
 import java.math.BigDecimal;
 import java.util.Map;
