+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -14,7 +14,7 @@ import kitchenpos.menu.infra.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.infra.ProductRepository;
 import kitchenpos.common.infra.PurgomalumClient;
-import kitchenpos.product.ui.dto.ProductCreationRequest;
+import kitchenpos.product.application.dto.ProductCreationRequest;
 
 @Service
 public class ProductService {
