+++ src/main/java/kitchenpos/product/application/dto/ProductCreationRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.ui.dto;
+package kitchenpos.product.application.dto;
 
 import java.math.BigDecimal;
 import java.util.UUID;
